3-CodigoAdicional/4x-resolução/requirements.txt
Pillow
RealESRGAN
streamlit
torch
